AVD-90 Add restore method to Athena API
